requirements.txt
streamlit
python-docx
beautifulsoup4
lxml